+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 * Add TypeScript type guards for each resource class
 * Add a Region mixin for Region typing
 * Add a DatabaseSlug mixin for Database Cluster size types
+* Add a DropletSlug mixin for Droplet size types
 
 ---